msm: mdss: enable partial update with ping-pong split
When ping-pong split is enabled, frame-buffer is not split and
user-land program is not aware that at lower level there are two
physical panels. Because of this, user-land will always send a single
ROI covering the entire frame-buffer. In such cases, it is driver's
responsibility to split the ROI into two when ROI is crossing mid-point
and make topology changes when ROI is only on left or right panel. Add
these changes to support partial update with ping-pong split.

Change-Id: I3fa1beac2f7e9700d5b3a3d4fda63078f1e6e976
Signed-off-by: Ujwal Patel <ujwalp@codeaurora.org>

Conflicts:
	drivers/video/msm/mdss/mdss_mdp_ctl.c